feat(commerce-enterprise): complete task group 8 — order 읽기 모델·장바구니·주문서·원 단위 금액
상품·판매자·쿠폰·사용자 쿠폰·포인트 읽기 모델(이벤트, 옛 이벤트 거르기), Money → Long 원(unit_price_won
확장·백필), 장바구니, 주문서(서버 가격·쿠폰 견적·포인트·판매자 배송비·수수료 스냅샷·안분 잔차 최대 라인,
만료 15분), 상품 재발행 어드민 API.

+++ order/domain/src/test/kotlin/com/kgd/order/domain/order/model/OrderTest.kt
@@ -3,7 +3,6 @@ package com.kgd.order.domain.order.model
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.BehaviorSpec
 import io.kotest.matchers.shouldBe
-import java.math.BigDecimal
 
 class OrderTest : BehaviorSpec({
     given("주문 생성 시") {
@@ -11,10 +10,10 @@ class OrderTest : BehaviorSpec({
             then("PENDING 상태의 주문이 생성되어야 한다") {
                 val order = Order.create(
                     userId = "user-1",
-                    items = listOf(OrderItem.of(productId = 1L, quantity = 2, unitPrice = Money(5000.toBigDecimal())))
+                    items = listOf(OrderItem.of(productId = 1L, quantity = 2, unitPrice = Money(5000L)))
                 )
                 order.status shouldBe OrderStatus.PENDING
-                order.totalAmount.amount shouldBe 10000.toBigDecimal()
+                order.totalAmount.amount shouldBe 10000L
                 order.userId shouldBe "user-1"
             }
         }
@@ -29,14 +28,14 @@ class OrderTest : BehaviorSpec({
     given("주문 완료 처리 시") {
         `when`("PENDING 상태이면") {
             then("COMPLETED로 전환되어야 한다") {
-                val order = Order.create("user-1", listOf(OrderItem.of(1L, 1, Money(1000.toBigDecimal()))))
+                val order = Order.create("user-1", listOf(OrderItem.of(1L, 1, Money(1000L))))
                 order.complete()
                 order.status shouldBe OrderStatus.COMPLETED
             }
         }
         `when`("PENDING이 아니면") {
             then("IllegalStateException이 발생해야 한다") {
-                val order = Order.create("user-1", listOf(OrderItem.of(1L, 1, Money(1000.toBigDecimal()))))
+                val order = Order.create("user-1", listOf(OrderItem.of(1L, 1, Money(1000L))))
                 order.complete()
                 shouldThrow<IllegalStateException> {
                     order.complete()
@@ -47,7 +46,7 @@ class OrderTest : BehaviorSpec({
     given("주문 취소 시") {
         `when`("PENDING 상태이면") {
             then("CANCELLED로 전환되어야 한다") {
-                val order = Order.create("user-1", listOf(OrderItem.of(1L, 1, Money(1000.toBigDecimal()))))
+                val order = Order.create("user-1", listOf(OrderItem.of(1L, 1, Money(1000L))))
                 order.cancel()
                 order.status shouldBe OrderStatus.CANCELLED
             }
@@ -57,14 +56,14 @@ class OrderTest : BehaviorSpec({
         `when`("수량이 0이면") {
             then("IllegalArgumentException이 발생해야 한다") {
                 shouldThrow<IllegalArgumentException> {
-                    OrderItem.of(1L, 0, Money(1000.toBigDecimal()))
+                    OrderItem.of(1L, 0, Money(1000L))
                 }
             }
         }
         `when`("소계 계산") {
             then("unitPrice * quantity가 반환되어야 한다") {
-                val item = OrderItem.of(1L, 3, Money(1000.toBigDecimal()))
-                item.subtotal.amount shouldBe 3000.toBigDecimal()
+                val item = OrderItem.of(1L, 3, Money(1000L))
+                item.subtotal.amount shouldBe 3000L
             }
         }
     }
