+++ order/feature/src/main/kotlin/com/kgd/order/application/order/usecase/GetOrderUseCase.kt
@@ -1,6 +1,5 @@
 package com.kgd.order.application.order.usecase
 
-import java.math.BigDecimal
 
 interface GetOrderUseCase {
     fun execute(id: Long): Result
@@ -8,7 +7,7 @@ interface GetOrderUseCase {
     data class Result(
         val orderId: Long,
         val userId: String,
-        val totalAmount: BigDecimal,
+        val totalAmount: Long,
         val status: String
     )
 }
