+++ order/feature/src/main/kotlin/com/kgd/order/presentation/order/dto/OrderResponse.kt
@@ -2,12 +2,11 @@ package com.kgd.order.presentation.order.dto
 
 import com.kgd.order.application.order.usecase.GetOrderUseCase
 import com.kgd.order.application.order.usecase.PlaceOrderUseCase
-import java.math.BigDecimal
 
 data class OrderResponse(
     val orderId: Long,
     val userId: String,
-    val totalAmount: BigDecimal,
+    val totalAmount: Long,
     val status: String
 ) {
     companion object {
